+++ types/meteor/session.d.ts
@@ -1,0 +1,22 @@
+
+declare module Session {
+    function equals(key: string, value: string | number | boolean | any): boolean;
+
+    function get(key: string): any;
+
+    function set(key: string, value: EJSONable | any): void;
+
+    function setDefault(key: string, value: EJSONable | any): void;
+}
+
+declare module "meteor/session" {
+    module Session {
+        function equals(key: string, value: string | number | boolean | any): boolean;
+
+        function get(key: string): any;
+
+        function set(key: string, value: EJSONable | any): void;
+
+        function setDefault(key: string, value: EJSONable | any): void;
+    }
+}
