+++ types/meteor/templating.d.ts
@@ -1,0 +1,16 @@
+
+declare var Template: TemplateStatic;
+interface TemplateStatic extends Blaze.TemplateStatic {
+    new (viewName?: string, renderFunction?: Function): Blaze.Template;
+    body: Blaze.Template;
+    [index: string]: any | Blaze.Template;
+}
+
+declare module "meteor/templating" {
+    var Template: TemplateStatic;
+    interface TemplateStatic extends Blaze.TemplateStatic {
+        new (viewName?: string, renderFunction?: Function): Blaze.Template;
+        body: Blaze.Template;
+        [index: string]: any | Blaze.Template;
+    }
+}
